requirements_dev.txt
aniso8601==7.0.0
atomicwrites==1.3.0
attrs==19.1.0
bcbio-gff==0.6.6
biopython==1.73
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
configparser==3.5.0
contextlib2==0.5.5
-e git+https://github.com/mutalyzer/crossmap.git@851eac3ab235d53aae298cae0e3398fc54da2b4b#egg=crossmapper
-e git+https://github.com/mutalyzer/description-extractor.git@2c2edbffa063749bb3894518d87e29fb26fafe2d#egg=description_extractor
flasgger==0.9.3
Flask==1.0.3
Flask-JSONRPC==0.3.1
flask-restx==0.1.1
future==0.17.1
idna==2.8
importlib-metadata==0.15
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
lark-parser==0.7.5
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.0.0
-e git+https://github.com/mutalyzer/hgvs-parser@a182a3680381de5c0f6da5384a9594b396db00d4#egg=mutalyzer_hgvs_parser
-e git+https://github.com/mutalyzer/mutator.git@ddbebb6d75726d0cf9f41a6a4d319a3ef42ff760#egg=mutalyzer_mutator
-e git+https://github.com/mutalyzer/normalizer.git@gluing_new_model#egg=normalizer
numpy==1.16.3
pathlib2==2.3.3
pluggy==0.12.0
py==1.8.0
pydot==1.4.1
pyparsing==2.2.0
pyrsistent==0.15.4
pytest==4.5.0
pytz==2019.2
PyYAML==5.1.2
redis==3.2.1
requests==2.21.0
-e git+https://github.com/mutalyzer/retriever.git@088c4ce397d18fc225c10bfae2bacf6f5246efb7#egg=retriever
schema==0.7.0
six==1.12.0
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
